chore: wip tests
