+++ src/components/course/course-header/course-run-actions/data/index.js
@@ -1,0 +1,1 @@
+export { default as messages } from './messages';
