+++ src/components/course/course-header/course-run-actions/data/messages.js
@@ -1,0 +1,11 @@
+import { defineMessages } from '@edx/frontend-platform/i18n';
+
+const messages = defineMessages({
+  viewCourse: {
+    id: 'useCourseRunCardAction.viewCourse',
+    defaultMessage: 'View course',
+    description: 'Label for button when learner is already enrolled.',
+  },
+});
+
+export default messages;
